requirements.txt
manim>=0.18.0
httpx>=0.27.0
python-dotenv>=1.0.0
